Share the existing WASM resource limits across host users
Move store construction, instruction and bulk metering, memory and table
limits, and their tests into sapio-wasm. Compiler plugins reexport the
same implementation; runtime behavior and limits are unchanged.

Keep the shared crate independent of the compiler, plugin ABI and signer
so evaluator execution can reuse its limits without those capabilities.

Validation: 10 shared runtime tests and 51 plugin host tests pass.
