+++ wasm/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+rust-version.workspace = true
+name = "sapio-wasm"
+version = "0.2.4"
+license = "MPL-2.0"
+edition = "2021"
+repository = "https://github.com/sapio-lang/sapio"
+description = "Shared deterministic WASM execution and guest bindings for Sapio"
+
+[features]
+default = []
+host = ["dep:wasmer", "dep:wasmer-middlewares", "dep:wasmer-types"]
+
+[dependencies]
+wasmer = { version = "6.1.0", optional = true }
+wasmer-middlewares = { version = "=6.1.0", optional = true }
+wasmer-types = { version = "=6.1.0", optional = true }
