+++ wasm/src/host/runtime/tests.rs
